feat: add vite bundler and pwa support
