+++ package.json
@@ -5,11 +5,11 @@
   "packageManager": "pnpm@10.28.0",
   "scripts": {
     "build": "run-s 'build:*'",
-    "build:remix": "remix build",
-    "build:server": "esbuild --platform=node --format=cjs ./server.ts --outdir=build --bundle --external:fsevents",
+    "build:remix": "vite build",
+    "build:server": "esbuild --platform=node --format=cjs ./server.ts --outdir=build --bundle --external:fsevents --external:vite",
     "dev": "run-p 'dev:*'",
     "dev:server": "cross-env NODE_ENV=development npm run build:server -- --watch",
-    "dev:remix": "remix dev --manual -c \"node ./build/server.js\"",
+    "dev:remix": "cross-env NODE_ENV=development node --watch-path=./build/server.js ./build/server.js",
     "start": "cross-env NODE_ENV=production node ./build/server.js",
     "format": "prettier --write .",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
@@ -51,8 +51,7 @@
     "@radix-ui/react-tooltip": "^1.1.2",
     "@react-email/components": "^0.0.14",
     "@react-email/render": "0.0.12",
-    "@remix-run/css-bundle": "^2.10.0",
-    "@remix-run/express": "^2.10.0",
+"@remix-run/express": "^2.10.0",
     "@remix-run/node": "^2.10.0",
     "@remix-run/react": "^2.10.0",
     "@remix-run/server-runtime": "^2.10.0",
@@ -148,6 +147,7 @@
     "tsx": "^4.19.4",
     "typescript": "^5.5.3",
     "vite": "^5.3.2",
+    "vite-plugin-pwa": "^1.2.0",
     "vite-tsconfig-paths": "^4.3.2",
     "vitest": "^1.6.0"
   },
